Update HistoryUtil::GetSignInState() to return an enum instead of a bool
This makes the code a little bit clearer, but more importantly, it sets
us up for introducing additional states soon.

All callers in both C++ and TypeScript are updated to use the new enum.

There should be no behavior changes.

AI notes: With very minor exceptions (like updating a TODO), the entire
CL was typed by Gemini CLI. This required quite a lot of iteration, and
overall took way longer than doing things manually. But hey, it works!

Bug: 417950948, 418144047
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6839141
Commit-Queue: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1500666}

+++ chrome/browser/ui/webui/history/history_ui.cc
@@ -266,7 +266,8 @@ void HistoryUI::UpdateDataSource() {
   Profile* profile = Profile::FromWebUI(web_ui());
 
   base::Value::Dict update;
-  update.Set(kSignInStateKey, HistoryUtil::GetSignInState(profile));
+  update.Set(kSignInStateKey,
+             static_cast<int>(HistoryUtil::GetSignInState(profile)));
 
   const bool is_managed = profile->GetPrefs()->IsManagedPreference(
       history_clusters::prefs::kVisible);
